Cast value to long instead of int, shouldn't break any sane system, fixes warning on 64bit systems
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@4841 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/inspircd.cpp
@@ -592,7 +592,7 @@ bool InspIRCd::LoadModule(const char* filename)
 				snprintf(MODERR,MAXBUF,"Loader/Linker error: %s",factory[MODCOUNT+1]->LastError());
 				return false;
 			}
-			if ((int)factory[MODCOUNT+1]->factory != -1)
+			if ((long)factory[MODCOUNT+1]->factory != -1)
 			{
 				Module* m = factory[MODCOUNT+1]->factory->CreateModule(MyServer);
 				modules[MODCOUNT+1] = m;
